feat(sys): 添加用户管理模块及相关功能
- 在sys模块中添加用户、部门、菜单和角色的控制器及服务
- 更新用户控制器，新增分页和下拉接口
- 在app模块中导入sys模块
- 导出BaseResult类以支持结果处理
- 修正sys-menu-entity.ts中的排序字段类型

+++ apps/admin/src/module/sys/_gen/_entity/sys-menu-entity.ts
@@ -33,6 +33,6 @@ export class SysMenuEntity extends BaseRecordDb {
 	@Column({ type: "varchar", length: 128, unique: true, comment: "权限标识" })
 	permission: string
 
-	@Column({ type: "int8", default: 0, comment: "排序" })
+	@Column({ type: "int", default: 0, comment: "排序" })
 	sort: number
 }
